requirements.txt
# alembic==1.7.6.dev0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
blinker==1.4
build==1.2.1
CacheControl==0.14.0
certifi==2024.6.2
cffi==1.17.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
command-not-found==0.3
crashtest==0.4.1
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.1.2
exceptiongroup==1.2.2
fastapi==0.112.0
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
fastjsonschema==2.20.0
filelock==3.15.3
greenlet==1.1.2
gyp==0.1
h11==0.14.0
httplib2==0.20.2
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.7.1
keyring==24.3.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
makefun==1.15.4
Mako==1.1.3
MarkupSafe==2.0.1
more-itertools==8.10.0
msgpack==1.0.3
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.1
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pwdlib==0.2.0
pycairo==1.20.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyGObject==3.42.1
PyJWT==2.8.0
pynvim==0.4.2
pyparsing==2.4.7
pyproject_hooks==1.1.0
pyTelegramBotAPI==4.21.0
pytest==8.3.2
python-apt==2.4.0+ubuntu3
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==5.4.1
rapidfuzz==3.9.3
requests==2.32.3
requests-toolbelt==1.0.0
SecretStorage==3.3.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
systemd-python==234
tomli==2.0.1
tomlkit==0.12.5
trove-classifiers==2024.5.22
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.26.2
wadllib==1.3.6
zipp==1.0.0
alembic==1.7.6.dev0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
beanie==1.26.0
blinker==1.4
build==1.2.1
CacheControl==0.14.0
certifi==2024.6.2
cffi==1.17.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
command-not-found==0.3
crashtest==0.4.1
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.1.2
exceptiongroup==1.2.2
fastapi==0.112.0
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
fastjsonschema==2.20.0
filelock==3.15.3
greenlet==1.1.2
gyp==0.1
h11==0.14.0
httplib2==0.20.2
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.7.1
keyring==24.3.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-model==0.2.0
makefun==1.15.4
Mako==1.1.3
MarkupSafe==2.0.1
more-itertools==8.10.0
motor==3.5.1
msgpack==1.0.3
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.1
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pwdlib==0.2.0
pycairo==1.20.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyGObject==3.42.1
PyJWT==2.8.0
pymongo==4.8.0
pynvim==0.4.2
pyparsing==2.4.7
pyproject_hooks==1.1.0
pyTelegramBotAPI==4.21.0
pytest==8.3.2
python-apt==2.4.0+ubuntu3
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==5.4.1
rapidfuzz==3.9.3
requests==2.32.3
requests-toolbelt==1.0.0
SecretStorage==3.3.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
systemd-python==234
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
trove-classifiers==2024.5.22
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.26.2
wadllib==1.3.6
zipp==1.0.0
